+++ test/tests/test_brillouin.jl
@@ -43,7 +43,7 @@ end
     # mapping to mesh index
     for trial in 1 : 10
         p  = rand(points(m))
-        n  = SVector{2, Int64}(rand(-4 : 4), rand(-4 : 4))
+        n  = SVector{2, Int}(rand(-4 : 4), rand(-4 : 4))
         q1 = value(p) + BrillouinPoint(domain(m)[:bz].L .* n)
         q2 = euclidean(p, m) + basis(domain(m)[:bz]) * n
